P-22_1_6 create patient list + creat design
