+++ catchBullying/Doctor/DoctorProfileDetailsViewController.swift
@@ -48,7 +48,7 @@ class DoctorProfileDetailsViewController: UIViewController {
   }
   
   func gotoChat(conversation: ConversationModel) {
-    let controller = self.storyboard?.instantiateViewController(identifier: "ChatVC") as! ChattViewController
+    let controller = self.storyboard?.instantiateViewController(identifier: "ChatVC") as! ChatViewController
     controller.conversation = conversation
     controller.modalPresentationStyle = .fullScreen
     controller.modalTransitionStyle = .flipHorizontal
