+++ catchBullying/ViewControllers/LandingViewController.swift
@@ -14,7 +14,7 @@ class LandingViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    try! Auth.auth().signOut()
+    //try! Auth.auth().signOut()
   }
   
   override func viewWillAppear(_ animated: Bool) {
